fix a bug in deleteReservation

+++ src/main/java/com/gatorRider/GatorRider/Service/RideService.java
@@ -193,15 +193,16 @@ public class RideService implements org.hibernate.service.Service {
 
         for(RidePassenger i: list){
             //notify passenger
-            notificationService.sendSMSToOne(driverRepository.getOne(ridePassenger.getPassengerId()),
-                    SMSMessage.NOTIFY_PASSENGER_WHEN_RIDERESERVE_IS_REMOVED
-            );
-            //update seat
-            ridePassengerRepository.deleteById(i.getId());
-            Ride tempRide = rideRepository.getOne(ridePassenger.getRideId());
-            tempRide.setNumSeatAvailable(tempRide.getNumSeatAvailable() + 1);
-            rideRepository.save(tempRide);
-            System.out.println(tempRide);
+            if(i.getPassengerId().equals(ridePassenger.getPassengerId())){
+                notificationService.sendSMSToOne(driverRepository.getOne(ridePassenger.getPassengerId()),
+                        SMSMessage.NOTIFY_PASSENGER_WHEN_RIDERESERVE_IS_REMOVED
+                );
+                //update seat
+                ridePassengerRepository.deleteById(i.getId());
+                Ride tempRide = rideRepository.getOne(ridePassenger.getRideId());
+                tempRide.setNumSeatAvailable(tempRide.getNumSeatAvailable() + 1);
+                rideRepository.save(tempRide);
+            }
 
         }
     }
